+++ vista/src/annotations/plugins/index.js
@@ -1,6 +1,7 @@
 // src/annotations/plugins/index.js
 
 // Import all the plugin definition files
+//import annotationListItemPlugin from "./annotationListItemPlugin";
 import annotationCreationCompanionWindow from "./annotationCreationCompanionWindow";
 import canvasAnnotationsPlugin from "./canvasAnnotationsPlugin";
 import externalStorageAnnotationPlugin from "./externalStorageAnnotationPlugin";
@@ -9,6 +10,7 @@ import windowSideBarButtonsPlugin from "./windowSideBarButtonsPlugin";
 
 // Export them together in an array
 const plugins = [
+  //annotationListItemPlugin,
   annotationCreationCompanionWindow,
   canvasAnnotationsPlugin,
   externalStorageAnnotationPlugin,